Refactor blinkist scraper into dedicated class
Since the cloudscraper is reused internally multiple times it makes sense to save it as an object property. This should also simplify the code base.
