+++ dailyblink/__main__.py
@@ -1,19 +1,12 @@
 import argparse
 import pathlib
-import time
-
-import cloudscraper
-from cloudscraper import CloudflareChallengeError
 
 from dailyblink import __version__
+from dailyblink.core import BlinkistScraper
 from dailyblink.settings import (
-    AVAILABLE_LANGUAGES,
     BLINKS_DEFAULT_PATH,
     BLINKS_DIR_NAME,
-    MAX_CLOUDFLARE_ATTEMPTS,
-    CLOUDFLARE_WAIT_TIME,
 )
-from dailyblink.utils import download_blinks
 
 
 def parse_arguments():
@@ -51,31 +44,10 @@ def parse_arguments():
 
 def main():
     args = parse_arguments()
-    print("Downloading the free daily Blinks...")
-
-    languages = {
-        key: value
-        for key, value in AVAILABLE_LANGUAGES.items()
-        if value in args.language
-    }
 
     base_path = pathlib.Path(args.path) / BLINKS_DIR_NAME
-    scraper = cloudscraper.create_scraper()
-
-    try:
-        download_blinks(scraper, languages, base_path)
-    except CloudflareChallengeError as e:
-        print(e)
-        print("Retrying...")
-        for index in range(1, MAX_CLOUDFLARE_ATTEMPTS + 1):
-            print(f"Attempt {index}/{MAX_CLOUDFLARE_ATTEMPTS}... ", end="")
-            time.sleep(CLOUDFLARE_WAIT_TIME)
-            scraper = cloudscraper.create_scraper()
-            try:
-                download_blinks(scraper, languages, base_path)
-                break
-            except CloudflareChallengeError:
-                print("FAILED")
+    blinkist_scraper = BlinkistScraper()
+    blinkist_scraper.download_daily_blinks(args.language, base_path)
 
 
 if __name__ == "__main__":
