+++ dailyblink/core.py
@@ -1,9 +1,163 @@
 import pathlib
+import re
+import time
+from datetime import date
 
+import cloudscraper
 from bs4 import BeautifulSoup
-from mutagen.mp4 import MP4
+from cloudscraper import CloudflareChallengeError
 
-from dailyblink.settings import BASE_URL, COVER_FILE_NAME
+from dailyblink.media import (
+    save_book_text,
+    set_m4a_meta_data,
+    save_media,
+)
+from dailyblink.settings import (
+    BASE_URL,
+    COVER_FILE_NAME,
+    PLAYLIST_FILE_NAME,
+    LANGUAGES,
+    MAX_CLOUDFLARE_ATTEMPTS,
+    CLOUDFLARE_WAIT_TIME,
+)
+
+
+class BlinkistScraper:
+    def __init__(self):
+        self._reset_cloudscraper()
+
+    def download_daily_blinks(self, languages: list, base_path: pathlib.Path):
+        print("Downloading the free daily Blinks...")
+        try:
+            self._attempt_daily_blinks_download(languages, base_path)
+        except CloudflareChallengeError as e:
+            print(e)
+            print("Retrying...")
+            for index in range(1, MAX_CLOUDFLARE_ATTEMPTS + 1):
+                print(f"Attempt {index}/{MAX_CLOUDFLARE_ATTEMPTS}... ", end="")
+                time.sleep(CLOUDFLARE_WAIT_TIME)
+                self._reset_cloudscraper()
+                try:
+                    self._attempt_daily_blinks_download(languages, base_path)
+                    break
+                except CloudflareChallengeError:
+                    print("FAILED")
+
+    def _reset_cloudscraper(self):
+        self.scraper = cloudscraper.create_scraper()
+
+    def _attempt_daily_blinks_download(self, languages: list, base_path: pathlib.Path):
+        for language_code in languages:
+            self._download_daily_blinks(language_code, base_path)
+
+        print()
+        print(f"All blinks were saved under {base_path.absolute()}")
+
+    def _download_daily_blinks(self, language_code, base_path):
+        language = LANGUAGES.get(language_code)
+        blink_info = self._get_daily_blink_info(language=language_code)
+        blink_url = blink_info["url"]
+
+        print()
+        print(f"{language} ({language_code}):")
+        print(f"{blink_info['title']} - {blink_info['author']}\n")
+
+        blink = self._request_blinkist_book_text(blink_url=blink_url)
+        book_id = blink["book-id"]
+        chapter_ids = blink["chapter-ids"]
+        chapters = blink["chapters"]
+
+        valid_title = re.sub(r"([^\s\w]|_)+", "", blink_info["title"])
+        valid_author = re.sub(r"([^\s\w]|_)+", "", blink_info["author"])
+        book_path = base_path / language / f"{date.today()} - {valid_title}"
+
+        print("Saving book text...")
+        save_book_text(
+            blink_info=blink_info,
+            chapters=chapters,
+            file_path=book_path / f"{valid_title} - {valid_author}.md",
+        )
+
+        print("Saving book cover...")
+        cover_response = self.scraper.get(blink_info["cover_url"])
+
+        save_media(
+            media=cover_response.content,
+            file_path=book_path / COVER_FILE_NAME,
+        )
+
+        try:
+            file_list = []
+            for number, chapter_id in enumerate(chapter_ids):
+                print(
+                    f"Saving audio track #{number + 1} - {chapters[number][0][:40]}..."
+                )
+                file_name = f"{number:02d} - {valid_title}.m4a"
+                file_path = book_path / file_name
+                file_list.append(file_name)
+                audio_response = self._request_audio(
+                    book_id=book_id, chapter_id=chapter_id
+                )
+                save_media(media=audio_response.content, file_path=file_path)
+                set_m4a_meta_data(
+                    filename=file_path,
+                    artist=blink_info["author"],
+                    title=chapters[number][0],
+                    album=blink_info["title"],
+                    track_number=number + 1,
+                    total_track_number=len(chapter_ids),
+                    genre="Blinkist audiobook",
+                )
+            with open(book_path / PLAYLIST_FILE_NAME, "w") as f:
+                print("Creating playlist file...")
+                f.write("\n".join(file_list))
+        except ValueError:
+            print("No audio tracks are available.")
+
+    def _get_daily_blink_info(self, language="en"):
+        daily_blink_url = f"{BASE_URL}/{language}/nc/daily/"
+        response = self.scraper.get(daily_blink_url)
+        return _create_blink_info(response.text)
+
+    def _request_blinkist_book_text(self, blink_url):
+        response = self.scraper.get(blink_url)
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        book_id = soup.find_all("div", {"class": "reader__container"})[0][
+            "data-book-id"
+        ]
+
+        chapter_list_elements = soup.find_all("li", {"class": "chapters"})
+        chapter_ids = [
+            chapter_element["data-chapterid"]
+            for chapter_element in chapter_list_elements
+            if "data-chapterid" in chapter_element.attrs
+        ]
+
+        chapters = soup.find_all(None, {"class": "chapter chapter"})
+        book_text = [chapter.text.strip() for chapter in chapters]
+
+        chapter_texts = [
+            (chapter.split("\n", 1)[0], chapter.split("\n", 1)[1])
+            for chapter in book_text
+        ]
+
+        return {
+            "book-id": book_id,
+            "chapter-ids": chapter_ids,
+            "chapters": chapter_texts,
+        }
+
+    def _request_audio(self, book_id, chapter_id):
+        url = f"{BASE_URL}/api/books/{book_id}/chapters/{chapter_id}/audio"
+        headers = {"x-requested-with": "XMLHttpRequest"}
+        response = self.scraper.get(url, headers=headers)
+        if response.status_code == 404:  # Text only book, no audio
+            raise ValueError(
+                f"Audio track does not exist for book {book_id} chapter {chapter_id}"
+            )
+        audio_url = response.json().get("url")
+        return self.scraper.get(audio_url)
 
 
 def _create_blink_info(response_text):
@@ -31,110 +185,3 @@ def _create_blink_info(response_text):
         "about_author": about_author,
         "cover_url": cover_url,
     }
-
-
-def get_daily_blink_info(scraper, language="en"):
-    daily_blink_url = f"{BASE_URL}/{language}/nc/daily/"
-    response = scraper.get(daily_blink_url)
-    return _create_blink_info(response.text)
-
-
-def request_blinkist_book_text(scraper, blink_url):
-    response = scraper.get(blink_url)
-    soup = BeautifulSoup(response.text, "html.parser")
-
-    book_id = soup.find_all("div", {"class": "reader__container"})[0]["data-book-id"]
-
-    chapter_list_elements = soup.find_all("li", {"class": "chapters"})
-    chapter_ids = [
-        chapter_element["data-chapterid"]
-        for chapter_element in chapter_list_elements
-        if "data-chapterid" in chapter_element.attrs
-    ]
-
-    chapters = soup.find_all(None, {"class": "chapter chapter"})
-    book_text = [chapter.text.strip() for chapter in chapters]
-
-    chapter_texts = [
-        (chapter.split("\n", 1)[0], chapter.split("\n", 1)[1]) for chapter in book_text
-    ]
-
-    return {"book-id": book_id, "chapter-ids": chapter_ids, "chapters": chapter_texts}
-
-
-def request_audio(scraper, book_id, chapter_id):
-    url = f"{BASE_URL}/api/books/{book_id}/chapters/{chapter_id}/audio"
-    headers = {"x-requested-with": "XMLHttpRequest"}
-    response = scraper.get(url, headers=headers)
-    if response.status_code == 404:  # Text only book, no audio
-        raise ValueError(
-            f"Audio track does not exist for book {book_id} chapter {chapter_id}"
-        )
-    audio_url = response.json().get("url")
-    return scraper.get(audio_url)
-
-
-def save_audio_content(audio_response, file_path):
-    pathlib.Path(file_path).parent.mkdir(parents=True, exist_ok=True)
-
-    with open(file_path, "wb") as file:
-        file.write(audio_response.content)
-
-
-def save_book_cover(scraper, cover_url, file_path):
-    pathlib.Path(file_path).parent.mkdir(parents=True, exist_ok=True)
-
-    with open(file_path, "wb") as file:
-        file.write(scraper.get(cover_url).content)
-
-
-def save_book_text(blink_info, chapters, file_path, cover_path=COVER_FILE_NAME):
-    pathlib.Path(file_path).parent.mkdir(parents=True, exist_ok=True)
-
-    with open(file_path, "w+") as file:
-        file.write(f"# {blink_info['title']}\n\n")
-        file.write(f"_{blink_info['author']}_\n\n")
-        file.write(f"{blink_info['read_time']}\n\n")
-        file.write(f"![cover]({cover_path})\n\n")
-
-        file.write(f"### Synopsis\n\n{blink_info['synopsis']}\n\n")
-        file.write(f"### Who is it for?\n\n{blink_info['for_who']}\n\n")
-        file.write(f"### About the author\n\n{blink_info['about_author']}\n\n")
-        for number, chapter in enumerate(chapters):
-            if number != 0 and number != len(chapters) - 1:
-                file.write(f"## Blink {number} - {chapter[0]}\n\n")
-            else:
-                file.write(f"## {chapter[0]}\n\n")
-
-            file.write(f"{chapter[1]}\n\n")
-
-        file.write(f"Source: {blink_info['url']}\n\n")
-
-
-def set_m4a_meta_data(
-    filename,
-    artist=None,
-    title=None,
-    album=None,
-    track_number=None,
-    total_track_number=None,
-    genre=None,
-):
-    mp4_file = MP4(filename)
-
-    if not mp4_file.tags:
-        mp4_file.add_tags()
-
-    tags = mp4_file.tags
-
-    if artist:
-        tags["\xa9ART"] = artist
-    if title:
-        tags["\xa9alb"] = album
-    if album:
-        tags["\xa9nam"] = title
-    if track_number and total_track_number:
-        tags["trkn"] = [(track_number, total_track_number)]
-    if genre:
-        tags["\xa9gen"] = genre
-    tags.save(filename)
